REFACTOR: moved Updade related stuff into NodeUpdates
